+++ language/runtime/src/world/lineage/image.rs
@@ -2,16 +2,16 @@ use serde::{Deserialize, Serialize};
 
 /// Image identifier for one materialized world state image.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-pub struct ImageId(u128);
+pub struct ImageId(u64);
 
 impl ImageId {
     /// Create a new image identifier.
-    pub const fn new(value: u128) -> Self {
+    pub const fn new(value: u64) -> Self {
         Self(value)
     }
 
     /// Return the raw image identifier value.
-    pub const fn get(self) -> u128 {
+    pub const fn get(self) -> u64 {
         self.0
     }
 }
